python-client/requirements.txt
grpcio~=1.27.1
pandas~=1.0.1